WorldRugbyDataConverter -> RankingsDataConverter & MatchesDataConverter
