+++ app/src/main/java/com/ricknout/rugbyranker/vo/MatchesDataConverter.kt
@@ -2,28 +2,11 @@ package com.ricknout.rugbyranker.vo
 
 import com.ricknout.rugbyranker.common.api.Match
 import com.ricknout.rugbyranker.common.api.WorldRugbyMatchesResponse
-import com.ricknout.rugbyranker.common.api.WorldRugbyRankingsResponse
 import com.ricknout.rugbyranker.common.api.WorldRugbyService
 import com.ricknout.rugbyranker.common.vo.Sport
 import java.lang.IllegalArgumentException
 
-object WorldRugbyDataConverter {
-
-    fun getWorldRugbyRankingsFromWorldRugbyRankingsResponse(worldRugbyRankingsResponse: WorldRugbyRankingsResponse, sport: Sport): List<WorldRugbyRanking> {
-        return worldRugbyRankingsResponse.entries.map { entry ->
-            WorldRugbyRanking(
-                    teamId = entry.team.id,
-                    teamName = entry.team.name,
-                    teamAbbreviation = entry.team.abbreviation!!,
-                    position = entry.pos,
-                    previousPosition = entry.previousPos,
-                    points = entry.pts,
-                    previousPoints = entry.previousPts,
-                    matches = entry.matches,
-                    sport = sport
-            )
-        }
-    }
+object MatchesDataConverter {
 
     fun getWorldRugbyMatchesFromWorldRugbyMatchesResponse(worldRugbyMatchesResponse: WorldRugbyMatchesResponse, sport: Sport): List<WorldRugbyMatch> {
         return worldRugbyMatchesResponse.content.map { match ->
